+++ docs/What is New/Changelog.md
@@ -15,8 +15,11 @@ _In recent [Tasks releases](https://github.com/obsidian-tasks-group/obsidian-tas
 - X.Y.Z:
   - [[Line Continuations]] can now be used in the [[Query File Defaults]] property `TQ_extra_instructions`.
   - [[Check your Statuses]] report now contains samples of each status, and a convenient search to test them.
+  - Add German translation of [[Settings]], [[Editing a Status]] and [[Check your Statuses]].
 - 7.19.0:
   - New setting to [[Recurring Tasks#Remove scheduled date on recurrence|remove scheduled date on recurrence]].
+- 7.17.0:
+  - Add Belarusian, Russian, and Ukrainian translations of [[Settings]], [[Editing a Status]] and [[Check your Statuses]].
 - 7.16.0:
   - Add `task.lineNumber`.
     - This enables `sort by function task.lineNumber` to override the [[Sorting#Default sort order|default sort order]].